--- src/app/layout.tsx
@@ -1,49 +1,0 @@
-import { ToastContainer } from "react-toastify";
-import { MangadexContextProvider } from "@/contexts/mangadex";
-import Gtag from "@/components/gtag";
-import { Metadata } from "next";
-import { Constants } from "@/constants";
-
-import "slick-carousel/slick/slick-theme.css";
-import "slick-carousel/slick/slick.css";
-import "react-toastify/dist/ReactToastify.css";
-import "swiper/css";
-import "@fortawesome/fontawesome-free/css/all.css";
-import "@/styles/base/index.scss";
-
-export const metadata: Metadata = {
-  title: `${Constants.APP_NAME} - Truyện tranh chất lượng cao không quảng cáo`,
-  description: `Đọc truyện miễn phí, chất lượng cao và tham gia ủng hộ nhóm dịch trên ${Constants.APP_NAME}`,
-  applicationName: Constants.APP_NAME,
-  authors: [{ name: "TruyenDex", url: "https://github.com/zennomi/truyendex" }],
-  keywords: [
-    "truyện tranh",
-    "manga",
-    "manhwa",
-    "manhua",
-    "nettruyen",
-    "nettrom",
-    "blogtruyen",
-    "truyendex",
-  ],
-  metadataBase: new URL(Constants.APP_URL),
-  other: {
-    referrer: "same-origin",
-  },
-};
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return (
-    <html lang="vi" className="dark">
-      <body>
-        <MangadexContextProvider>{children}</MangadexContextProvider>
-        <ToastContainer />
-        <Gtag />
-      </body>
-    </html>
-  );
-}
